Introduces `CustomSnackbar` for displaying stylized messages
This commit adds a new `CustomSnackbar` composable and its associated `CustomSnackBarVisuals` data class. This allows for displaying snackbars with custom container and content colors, and an optional leading icon.
